+++ src/Kentico.Xperience.Lucene.Core/Indexing/DefaultLuceneClient.cs
@@ -55,7 +55,7 @@ internal class DefaultLuceneClient : ILuceneClient
     private readonly IWebFarmService webFarmService;
 
 
-    private readonly LuceneIndexSearcherProvider searcherProvider;
+    private readonly LuceneSearchCacheInvalidator searchCacheInvalidator;
 
 
     private readonly LuceneSearchOptions luceneSearchOptions;
@@ -75,7 +75,7 @@ internal class DefaultLuceneClient : ILuceneClient
         IEventLogService log,
         ILuceneIndexManager indexManager,
         IWebFarmService webFarmService,
-        LuceneIndexSearcherProvider searcherProvider,
+        LuceneSearchCacheInvalidator searchCacheInvalidator,
         IOptions<LuceneSearchOptions> luceneSearchOptions
         )
     {
@@ -89,7 +89,7 @@ internal class DefaultLuceneClient : ILuceneClient
         this.log = log;
         this.indexManager = indexManager;
         this.webFarmService = webFarmService;
-        this.searcherProvider = searcherProvider;
+        this.searchCacheInvalidator = searchCacheInvalidator;
         this.luceneSearchOptions = luceneSearchOptions.Value;
     }
 
@@ -171,30 +171,9 @@ internal class DefaultLuceneClient : ILuceneClient
 
         bool result = await luceneIndex.StorageContext.DeleteIndex();
 
-        InvalidateSearchCache(luceneIndex);
+        searchCacheInvalidator.Invalidate(luceneIndex);
 
         return result;
-    }
-
-
-    /// <summary>
-    /// Invalidates the local cached searcher for the index and, when the index lives on external (shared)
-    /// storage, broadcasts the invalidation to the other web farm servers (where the reset/delete operation
-    /// itself is not replicated). For non-external storage the operation is replicated to each server via its
-    /// own web farm task, which invalidates that server's cache locally.
-    /// </summary>
-    private void InvalidateSearchCache(LuceneIndex luceneIndex)
-    {
-        searcherProvider.Invalidate(luceneIndex.IndexName);
-
-        if (StorageHelper.IsExternalStorage(luceneIndex.StorageContext.IndexStoragePathRoot))
-        {
-            webFarmService.CreateTask(new InvalidateSearchIndexWebFarmTask
-            {
-                IndexName = luceneIndex.IndexName,
-                CreatorName = webFarmService.ServerName
-            });
-        }
     }
 
     private Task<int> DeleteRecordsInternal(IEnumerable<string> itemGuids, string indexName)
@@ -235,7 +214,7 @@ internal class DefaultLuceneClient : ILuceneClient
 
         // The reset replaces the published generation (and retention may move the old one to .trash),
         // so drop any cached searcher pointing at it.
-        InvalidateSearchCache(luceneIndex);
+        searchCacheInvalidator.Invalidate(luceneIndex);
 
         var contentQueryExecutionOptions = new ContentQueryExecutionOptions
         {
